refactor: comp for toolbar, save modal
